scanner/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.34
botocore==1.38.34
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
jmespath==1.0.1
mysql-connector-python==9.3.0
numpy==2.2.6
pandas==2.3.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
